change the code for createPartner
